add: api 연결, collapsingtoolbar 직접 구현

+++ presentation/src/main/java/com/project/ibooku/presentation/ui/feature/review/screen/BookReviewWriteScreen.kt
@@ -79,7 +79,7 @@ fun BookReviewWriteScreen(
                         viewModel.onEvent(BookReviewEvents.SpoilerChanged(isSpoiler))
                     },
                     onRatingChanged = { rating ->
-                        viewModel.onEvent(BookReviewEvents.ReviewRatingChanged(rating))
+                        viewModel.onEvent(BookReviewEvents.ReviewRatingChanged(rating.toDouble()))
                     }
                 )
 
@@ -104,7 +104,7 @@ fun BookReviewWriteScreen(
 @Composable
 fun BookReviewWriteScreenBody(
     review: String,
-    rating: Float,
+    rating: Double,
     isSpoiler: Boolean?,
     modifier: Modifier = Modifier,
     onReviewTextChanged: (String) -> Unit,
@@ -130,7 +130,7 @@ fun BookReviewWriteScreenBody(
 
         RatingBar(
             modifier = Modifier.padding(horizontal = 22.dp),
-            value = rating,
+            value = rating.toFloat(),
             size = 50.dp,
             painterEmpty = painterResource(id = R.drawable.ic_review_star_empty),
             painterFilled = painterResource(id = R.drawable.ic_review_start_filled),
